+++ main.cpp
@@ -1,7 +1,7 @@
 /*
     Project/Script: ESP8266-01S Basic Connection
-    Version: 1.0
-    Hardware: NUCLEO L152RE mit Arduino Multifunction Shield on top of it.
+    Version: 1.1
+    Hardware: NUCLEO L152RE und ESP826-01S
 */
 #include "mbed.h"
 
@@ -14,9 +14,9 @@ TCPSocket tcp_socket;
 
 #define SLEEP_TIME 100ms
 
-DigitalIn t1(PA_1);
-DigitalIn t2(PA_4);
+DigitalIn t1(BUTTON1);
 const int PRESSED = 0;
+bool wlan_connected = false;
 
 // development/ debugging
 #define LOG_DEBUG print_message_to_serial
@@ -81,23 +81,25 @@ int disconnect_from_wifi() {
 int main()
 {
 
-    LOG_DEBUG("Ready, press Button 1 on MF-Shield to connect to WIFI");
+    LOG_DEBUG("Ready, press Blue Button L152RE to connect to WIFI");
 
     while (true) {
 
         if(t1 == PRESSED){
             while(t1 == PRESSED){}
-            LOG_DEBUG("Button 1 pressed! will now try to connect to WIFI");
-            if( connect_to_wifi() == 0) {
-                LOG_DEBUG("You can Press Button 2 to disconnect from WIFI");
-            }
-        }
 
-        if(t2 == PRESSED){
-            while(t2 == PRESSED){}
-            LOG_DEBUG("Button 2 pressed! will now try to disconnect from WIFI");
-            disconnect_from_wifi();
-        }
+            if(wlan_connected) { // wenn wlan ist verbunden, dann die verbindung trennen
+                LOG_DEBUG("Button pressed! will now try to disconnect from WIFI");
+                disconnect_from_wifi();
+                wlan_connected = !wlan_connected;
+            } else { // wenn wlan ist nicht verbunden, versuche die Verbindung herzustellen
+                LOG_DEBUG("Button pressed! will now try to connect to WIFI");
+                if( connect_to_wifi() == 0) { // wenn Verbindung erfolgreich war, die Meldung anzeigen
+                    wlan_connected = !wlan_connected;
+                    LOG_DEBUG("You can Press the Button again to disconnect from WIFI");
+                }
+            } // if(wlan_connected)
+        } // if(t1 == PRESSED)
 
         ThisThread::sleep_for(SLEEP_TIME);
     }