+++ changelog.md
@@ -1,3 +1,8 @@
 v1.0
     Basic WLAN Verbindungstest.
     NUCLEO-L152RE mit Arduino Multifunction Shield darauf. Die Tasten von MF-Shield werden für die Steuerung verwendet.
+
+v1.1
+    Basic WLAN Verbindungstest. (selbe wie v1.0, hier wird es nur ohne MF-Shield implementiert)
+    NUCLEO-L152RE ohne MF-Shield. Der Blau-Button on NUCLEO steuert das funktion
+